Made more sense to put switch in switch.py
